:recycle: split auth app tests
